Enable JVM IR backend
